Add crud operations for notes and users

+++ apps/server/src/notes/notes.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import { Injectable, NotFoundException, UnauthorizedException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { Note } from "./note.entity";
@@ -8,28 +8,38 @@ import { CreateNoteDto } from "./dtos/create-note.dto";
 export class NotesService {
   constructor(@InjectRepository(Note) private notesRepository: Repository<Note>) {}
 
-  findAll() {
-    return this.notesRepository.find();
+  findAll(userId: string) {
+    return this.notesRepository.findBy({ userId });
   }
 
-  findOne(id: string) {
-    return this.notesRepository.findOneBy({ id });
+  async findOne(noteId: string, userId: string) {
+    const note = await this.notesRepository.findOneBy({ id: noteId });
+
+    if (!note) throw new NotFoundException();
+
+    if (note.userId != userId) throw new UnauthorizedException();
+
+    return note;
   }
 
-  create(body: CreateNoteDto) {
+  create(body: CreateNoteDto, userId: string) {
     const { createdAt, title, text } = body;
     const newNote = this.notesRepository.create({
       title,
       text,
       createdAt,
+      userId,
     });
 
     return this.notesRepository.save(newNote);
   }
 
-  async delete(id: string) {
-    const note = await this.notesRepository.findOneBy({ id });
+  async delete(noteId: string, userId: string) {
+    const note = await this.notesRepository.findOneBy({ id: noteId });
+
     if (!note) throw new NotFoundException();
+
+    if (note.userId !== userId) throw new UnauthorizedException();
 
     return this.notesRepository.remove(note);
   }
